improves TabletServerGivesUpIT (#4290)
TabletServerGivesUpIT tests that a tserver will eventually die when it
can not write to a walog.  For walog write failures to occur, there must
be some thread in the tserver trying to write to the tserver. Changes in
elasticity have made it where the activitiy this test was doing did not
always cause a walog write.  This change is an attempt to improve the
chance of a tserver thread attempting a walog write after DFS is killed.