fixes clearing suspension for offline tables
There was code in the manager that seemed to have the intent of clearing
suspension markers in tablets for an offline table.  This code was not
working or tested.  This commit fixes this code and adds a tests that
validates that suspension markers are removed when a table is taken
offline.

fixes #3314