Created new ThriftClientType for shell authentication
TServerClient.getTabletServerConnection gets all of the
TabletServers and their ServiceLocks from ZooKeeper
via ZooCache. After that, it randomly gets a TabletServer
connection and returns it to the caller. When this is
done from the Shell it causes the initial login to wait
as all of this information is gathered. The information
gathering is useful as it populates ZooCache, but it can
take a long time on large clusters.

This change:

1. Adds a new ThriftClientType which overrides
getTabletServerConnection to quickly return a connection
to a random TabletServer
2. Adds Shell.initialAuthentication which uses the new
ThriftClientType to call authenticate user on the Client
service
3. Creates a background thread which calls the ping method
on a random TabletServer using the existing TServerClient
getTabletServerConnection to populate ZooCache.

Fixes #4303