fetches all columns in tablet mgmt iterator (#4307)
The tablet mgmt iterator was configuring a scanner to fetch almost all
of the tablets columns.  Since most columns were being fetched the
filtering offered no benefit and probably added extra cost.  It also
added code complexity.  This commit removes the fetch columns calls on
the scanner so that everything is fetched.