Considers all tablet metadata columns in split code (#4323)
Made the following changes to the split code that adds new tablets and
updates the existing tablet.

 * fixed potential NPE w/ tablet operation id check by reversing order
   of equals check
 * Throws IllegalStateException when attempting to split tablet with
   merged or cloned markers
 * Removed adding wals when creating new tablets in split, its not
   expected that the parent tablet would have wals and this is checked earlier
 * Deleted any user compaction requested, hosting requested, suspended,
   or last columns in the parent tablet

Added a unit test that attempts to exercise the split code with all
tablet columns.  The unit test also has a set of tablet columns that
were verified to work with split and it is checked against the set of
columns in the code.  The purpose of this test is to fail when a new
column is added to ensure that split is considered.

Was a bit uncertain about deleting the last location and suspend. Those
columns either need to be deleted from the parent tablet or added to the
new tablets being created.  The current code was doing neither.  Decided
to delete them as the new tablets have a different range and are
conceptually different tablets than the parent.