fixes bug and removes unneeded wait in delete table (#4346)
Noticed an integration test was taking 0 to 5 seconds to delete a table.
Investigated this and found some code in delete table that was waiting
on the TabletGroupWatcher to cycle as the cause of the delay.

Investigated why this wait was there and could not find a definite
answer.  Speculating that the purpose of the wait was to deal with
possible concurrent writes to tablet metadata by the tablet group
watcher in the case it has not yet noticed the table state is DELETING.
Changes in elasticity make this wait unnecessary because the delete table
fate sets an operation id on each tablet that should prevent any writes
to the tablet after its set.  After being set this would prevent any
in flight writes by the tablet group watcher from succeeding. Before
elasticity there was no way for the delete table fate to prevent writes
by the tablet group watcher, so thinking it just waited until it was
certain the tablet group watcher was aware of the new table state and
would therefore do no writes.

While researching I noticed the delete table code in 2.1 logged per
tablet information about any tablet its was waiting on.  This was not
being done in elasticity, so added logging for this.

With these changes TabletManagementIteratorIT went from 1 min runtime to
30 second runtime.  The test deleted 9 tables.