Updates merge code to handle all tablet metadata columns
Adds tests for all tablet metadata columns to ensure the merge code
handles them.  Some columns were not handled, so new code was added to
handle them.

In the process of adding unit test for the merge code, also added test
for some odd conditions in the code that would be hard to test in an
integration test.