Stop waiting on fate when reporting compaction complete (#4339)
Compactors were waiting on the Fate transaction that does compaction
commit.  They were doing this so that the dead compaction detector
would not kill the compaction during commit.  This change removes
the need for compactors to wait on compaction commit, freeing them
to run another compaction.

To accomplish this the dead compaction detector was modified to look
for fate operations commiting compactions.  This work leveraged the
new FateKey.  A bonus for this changes is that duplicate RPC messages
reporting a compaction as completed will no longer spin up duplicate
Fate operations to commit.  The Fate operations would have likely
been harmless, but would have consumed resources.

A new method was added to Fate to list transactions with a certain
FateKeyType.  Tests were added for this new Fate method.

An integration test was added to ensure the dead compaction detector
does not delete compactions that no compactors knows of but do have
a Fate transaction committing them.