Update TabletsMetadata stream() to autoclose TabletsMetadata (#4356)
Follow on changes to #4355 

Before this change if stream() was used the the TabletsMetadata close
method still had to be called manually as the close method on the stream
did not delegate. With this change if the stream is used in a
try-with-resources to autoClsoe the stream or if the stream is closed
manually, TabletsMetadata will also be closed.