partially avoids splitting and compacting offline tablets (#4343)
Modifies TabletManagementIterator so that it does not return tablets
that need split of compaction of if the table is offline.  This
partially prevents those tablets from splitting.  Still need to handle
things that were queued for split of compaction when a table is taken
offline.  Once those other changes are made this change will still avoid
queuing and processing tablets for work that does not need to be done.