Use fluent-style builder for pool creation, replacing overloaded methods
* Replaces overloaded createThreadPool methods with a fluent-style builder.
* Adds ThreadPoolBuilder test,
* Adds `createExecutorService` method that does not have an option to enable
metrics and replaces all occurrences in the code that was calling the alternate
method with emitThreadPoolMetrics=false to use it.

The `createScheduledExecutorService` will be refactored in a future PR when service
initialization is reworked.  See PR #4342 for an example.

This change isolates where metrics can be enabled and makes finding them
easier.  It also will make reviewing future changes easier becuase those
changes will be isolated to places that currently emable thead pool
metrics.