Create a test Ample implementation
This is a work in progress to create a test Ample implementation that
can be used to interact with Ample against a different table than the
table that the Accumulo instance uses. This will allow testing Ample
without having to deal with metadata being written to the table by the
system.

The test Ample also will support injecting custom behavior such as
changing the return status of a conditional mutation in order to test.

This closes #4376