cleaned up a few TODOs in the split code (#4401)
Removed TODOs in the code related to #3415, #3709, #3412.  These todos
were done or had issues opened.

Added some best effot checks to avoid doing splits if a table is
offline. The checks have race conditions as the table lock is not
acquired by the split fate op.  However as outlined in a comment
on #3412 the offline code that waits could be made more
comprehensive to account for this.

One TODO was handled by using
`Ample.RejectionHandler.acceptAbsentTablet()` which likely did not
exists when the TODO was written.