Logged msg instead of throwing exception in setFutureLocations
AbstractTabletStateStore.setFutureLocations was throwing an exception
when it could not set a future location in a tablets metadata. However,
it's possible that a concurrent FATE operation could prevent this and
this is not an error conditition. Modified the code to log a msg instead
as TabletGroupWatcher in the Manager will retry to assign the tablet when
the FATE operation has concluded and assignment is still needed.