fixes bug with starting multiple group of scan servers
When trying to start and stop multiple groups of scan servers using the
accumulo-cluster script the groups would interfere with each other.
This change fixes that by making the the scan servers use the group in
the instance id like compactors do.

Ran into this while workiing on #4444 and pulled it out as a stand alone
fix.