fixes conditional mutation bug when loading tablet
In #4436 a tablet update was moved to use conditional mutations. The
changed checked tablets current location in the conditional update.
However when this code was called in the Tablet constructor the tablet
server was the future location.  Adjusted to code to check for the
future location when calling from the constructor.