fixes continue scan in tablet mgmt iterator
Many places in the accumulo code will read a batch of key/values and
then use the last key in the batch to construct a range to get the next
batch.  The last key in the batch will be a non inclusive start key for
the range.  The tablet mgmt iterator was not handling this case and
returning the key that should have been excluded.