WIP add reservation related metrics to scan server
Scan servers differ from in tablet servers in that they can service
scans for any tablet, but must read the tablets files and reserve those
files to prevent the Accumulo GC from removing them.  This happens the
first time a tablet is read on a scan server and involves reads and
writes to the metadata table.  This PR is a start at adding metrics
related to reservations so that is possible to have visibility into the
impact of this process on scans.