batches reading tablet metadata for refresh (#4439)
Changes tablet server code that refreshes tablet metadata to read tablet
metadata for multiple tablets in a batch.  Also noticed the refresh code
was not handling some race condition correctly and could result in
spurious error messages.  It was also not handling closed tablets.
Refactored the code to return false when refresh is not possble for
acceptable reasons instead of throw an error.  This will cause the the
RPC to retry for those tablets later.  Also remove some code that was
attempting to handle complex race conditions that would be hard to test,
realized if these conditions do happen that it would be best to retry
rather than try to handle them.


Co-authored-by: Dave Marion <dlmarion@apache.org>