lowers logging level for first seen dead compaction
For the case of really quick compactions there is a good chance these
finish during the check done by the dead compaction detector.  Currently
when this happens a log message is logged about a possible dead compaction
that is a false positive.  A large number of quick external compactions
can cause a lot of these false positives.

This change adjust the first time a possible dead compaction is logged
to trace.