Modified Tablet Mutator implementations to always update srv:lock (#4421)
Modified the tablet mutating implementations (TabletMutator,
ConditionalTabletMutator, etc.) to always update the srv:lock
column so that the MetadataConstraints filter on the server side
validates that the update comes from a "valid" server. I had to
modify MiniAccumuloClusterImpl to acquire a ServiceLock for the ITs to
use where the test code is updating the tablet metadata. The changes
in the other classes are mainly just making sure that the tablet
mutator classes are getting the ServiceLock.

Closes #4420