Remove todo from SplitUtils (#4484)
I don't know that there is a reliable way to determine what the splits should be given the information that we have in the tablet metadata. The todo suggested running tests and doing some compactions, etc. But I think that it's really going to be situation dependent. Users can apply iterators to perform aggregation, deletes, etc. at compaction time that could influence the split points greatly. I think the better option here is to document compactions should be run first to get more accurate split points.