reduce the frequency that the list of scan servers is sorted
The provided plugin for working with scan servers sorts the list of scan
servers fairly frequently.  This list will not change too often, so it
could be sorted less to offer a better amortized cost as the total
number of scan servers grows.

Also outside of the plugin, the plugin was being reset every 100ms.
Removed this to give the plugin more control.  Also there is currently
no way to change client config for a created Accumulo client, so there
is no need to recreate the plugin.