Allow system compactions to run if zero user compaction jobs have run (#4480)
This change will allow system compactions to postpone user compactions
that have had no jobs run yet. Before this, if a user compaction was in
the queue and had selected files that overlapped it would block system
compactions from running. Now if there are selected files, but the user
compaction is not running and hasn't had any jobs completed, the
coordinator will clear the selectedFiles column so that the system
compaction can run if the expiration time has passed. The fate operation 
will reset the column again while trying to make progress.

This closes #4454

Co-authored-by: Keith Turner <kturner@apache.org>