fixes timeout in CompactionPriorityQueueMetricsIT
CompactionPriorityQueueMetricsIT was waiting up to 30s
for a dead compactor process to be absent in zookeeper.
The default zookeeper timeout is 30s, so there was a good
chance it would timeout.  Modified the zookeeper timeout
in the test to 10s rather then increase the time to wait
for the ZK entry to be absent.