lowers time to host ondemand tablets (#4581)
This change lowers the time it takes to host ondemand tablets
by moving this functionality into TabletGroupWatcher.

The client RPC thread processing the hosting request can now
directly call a function in TGW that will immediately start on
the work of hosting the tablets.

Updated SplitMillionIT to request hosting of 200 tablets all
at once instead of one by one.  This was done by using a
BatchScanner instead of lots of scanners.