handles failure in compaction planner
Updates the CompactionJobGenerator to log failures in planning compactions.
This avoids causing problems for the tablet management iterator when its
trying to do other things like process tablet location updates.

Added a test that ensures tables with flaky compaction planners can still
be read and written. If failed compaction planning were to interfere with
tablet assignment this would cause the reads and/or writes to fail.