improves fate execution of in progress transactions (#4589)
There are a few improvements to FATE in this PR.  The first
is that Fate now continues to execute a transaction as long
as its ready. Before it would do one step of a transaction
and then persist the result and stop, meaning the next stop
would not run until the next full scan.

The second change was to make finding runnable transactions
look for in progress transaction first in the persisted store
followed by the non inprogress ones.  This change causes the
search for runnable transactions to scan the store twice. To
make this more efficient a locality group and filtering iterator
were added.  This should make the scans looking for fate
transactions with a certain status much more efficient.

These two improvements should make fate a bit more responsive
for getting existing work done.

Co-authored-by: Dave Marion <dlmarion@apache.org>