ensure only single location is set in conditional location check (#4620)
Accumulo's metadata schema for locations allows multiple locations
to be set for a tablet even though only one is expected.  The
conditional mutation check for a location would succeed if the
required location was present, even if other locations were also
present.  Updated the check to make it more strict and only pass
if the required location is present AND no other locations are
present.

This update removes a specialized iterator for locations and uses
the more general SetEncodingIterator which allows treating a column
family like a map for comparison purposes.  Using this we can easily
check the more strict requirements using well tested code.

Added to ITs related to location conditional check to exercise the
situation of multiple locations being set on a tablet.