Modified FileCompactor to set interruptFlag on RFile
Added interrupt method on FileCompactor that is set to true
in CompactableUtils when the majc env is no longer enabled.
FileCompactor.interrupt sets an AtomicBoolean to true, and
the reference to the AtomicBoolean is set on the RFiles. Only
internal compactions will use this interrupt behavior as
external compactions already interrupt the compaction thread
when compaction cancellation is requested.

Fixes #4485