tests merge idempotence and fixes found bug (#4643)
Adds a way to make Fate in the manager always runs steps multiple times to test
idempotence. Created versions of merge and delete rows ITs that uses this test
manager. Found and fixed a bug in the merge code when running these new test.
Changed mini accumulo to allow specifying a different manager class to use prio
to startup.  There was an existing method that would allow starting a manager
with a different class, but using it would have meant letting the actuall
manager start and then killing it and starting another which would have added
time to test.  The change to mini accumulo was not made to its public API, only
its implementation.

fixes #4642