runs comprehensiveIT w/ FlakyFate and fixes found bug
Updated ComprehensiveIT to use FlakyFate because it calls a lot of
Fate operations. In the process found a bug with import table
where it failed when run a second time.  Import table had code
to handle running a second time but it compared filenames to full
file paths.  Converted everything in the code to use full file
paths to fix the bug.

Updated FlakyFate to interleave isReady() and call() in when rerunning
them as this is what would happen with a fault.