fixes ManagerAssignmentIT test failure
ManagerAssignmentIT had a test that would manually set an operation
id on a tablet and then attempt to set the tablet availibility on
the tablet using public API.  The changes in #4636 caused this public
API call to block because an opid was present.  Modified the test to
directly set tablet availability in the API.  Modifed the fate operation
to log the opid that is blocking it.