updates WriteAfterCloseIT to also use ConditionalWriter
Updated WriteAfterCloseIT to use a conditional writer in addition to
a batch writer.  Made a few improvements to the test.

While reviewing how the conditional writer handles timed out writes
noticed the tablet server code had an incorrect try/finally block.
There was code between session reservation and the beginning of
the try block.  If this inbetween code threw an exception, then it
would never unreserve the session.  Fixed this and added a few
comments to the code.

fixes #3742