fixes ExternalCompactionProgressIT
This test was validating the compactor busy count metric.  The test has 9
compactor processes running.  When it saw a busy count from any of the 9 it
would set an atomic long.  This is made it likely that the 8 of 9 not busy
compactors would set zero making the test flaky.  Replaced the atomic long w/ a
concurrent map where each compactor process has an entry in the map for its
busy count.