removes code that automatically removes multiple tablet locations
Code was added for [ACCUMULO-2261][1] that would automatically remove
duplication locations from tablet metadata.  This was an expected
event in the past because a tablet server could write a location
after it lost its lock and before the process died.

Now that conditional mutations are used to write locations and
with the changes in #4620 to make this really strict, the specific
problem described in ACCUMULO-2261 should no longer happen.
The code that was added to cleanup the metadata table for ACCUMULO-2261
could hide a new or different bug and prevent it from being found
and fixed. If a new cause of duplicate locations is found in the
elasticity branch then a fix that is appropriate for it should be
created.

Code was added to log detailed information when a duplicate location
is seen.  As part of this the root tablet metadata needed to be read
and that code was all over the place, so that was consolidated.

fixes #3888

[1]:https://issues.apache.org/jira/browse/ACCUMULO-2261