Tests UNKNOWN return status of conditional mutations (#4659)
These changes make it possible to instrument the manager
and tablet server in integration tests to use TestAmple
configured to randomly return UNKNOWN status for conditional
mutations.

Servers were instrumented by modifying the AbstractServer
constructor to accept a server context factory.  This
allowed test version of Manager and Tsever to pass a
factory that creates a server context that returns TestAmple.

A version of Comprehensive IT was updated to use the new
instrumented servers.  This should provide some base coverage
of UNKNOWN handling.  Still need to determine what Ample submit()
calls are not being exercised and modify more ITs to use
these instrumented servers.