Moves reading root tablet metadata w/ zoocache out of Ample (#4673)
After the changes in #4669 only a single location in the code was
reading tablet metadata w/ eventual consistency using Ample.  This
single location was the root tablet location cache.  Since this
was the only use of the code and its a highly nuanced use case,
moved the code out of Ample to the only place that was using it.
Removing the levels of indirection makes it easier to understand
the root tablet location cache. Also added some comments about
what the the root tablet location cache is doing and noticed and
uneeded cache clear it was doing and removed it.