adds auto split test to ComprehensiveIT (#4711)
ComprehensiveIT was covering all fate table op steps except
for the one related to automatic splits. This commit adds
a test to ComprehensiveIT that exercise autosplits.  This
covers that fate step that is only run for automatic splits.
Automatic splits are an important feature of Accumulo, so in
general its good to cover them in ComprehensiveIT.