fixes multiple upgrade bugs found when testing (#4717)
Fixed a bug where upgrade code was attempting to read the managers
service lock before it was set.

Fixed a bug with detecting fate operations.  The code to detect fate
operations would hang trying to read the fate table which did not
exists yet.

Scans of the metadata table running as the !SYSTEM user would fail
if no auths were specified.  The !SYSTEM user has not auths set in
zookeeper and an exception would happen when trying to find them.

A new zookeeper node related to compaction changes needed to be
created.

Simplified code for setting tablet availability by using DataLevel.


Co-authored-by: Dave Marion <dlmarion@apache.org>