Adds completable futures to compaction queue (#4726)
Adds completeable futures to the queue of compaction jobs.  This allows for
async notification when something is added to the queue.

The compaction queues code would drop queues that became empty.  The concept of
queues being empty became more complex with this change. A queue would be
considered empty when there were no futures and the queue was empty.  This
increased complexity of empty would have made the code for dropping empty
queues more complex.  Instead of increasing the complexity of this code chose
to drop removing empty queues.  This means that if a compaction group is used
and then no longer used that it will have a small empty datastructure sitting
around in map for the process lifetime.  That is unlikely to cause memory
issues.  Therefore decided the increased complexity was not worthwhile given
it was unlikely to cause memory problems.