Throw exception from  VisibilityFilter if thread interrupted
SystemIteratorUtil.setupSystemScanIterators puts the VisibilityFilter
at the top of the system scan iterator stack. This change modifies the
VisibilityFilter next, seek, and hasTop methods to check if the thread
running the scan has been interrupted. If it has, then these methods
now throw an InterationInterruptedException. This does not fully
address the situation mentioned in #4756, but it might help to address
user scans that do not check the Threads interrupted state when
processing Key/Values.

Related to #4756