tests log recovery using FlakyAmple
Modfied a log recovery test to run using FlakyAmple
This will provide coverage for the case of flay
conditional mutations that are written to tablets
as part log recovery.

Most of the code in WalSunnyDayIT was moved to
WalSunnyDatBaseIT w/o modification.  So even
though this commit has lots of diff, there are
very little actual changes.

This change is for #4710