Convert CompactionCoordinator RPC service to Async Thrift
This changes the CompactionCoordinator RPC service to use an Async
thrift processor. This allows long polling when a compactor requests a
new job by no longer blocking the Thrift IO thread when there are no
jobs available. When a job is available, the response will be sent
asynchronously back to the client.

The only RPC service converted for now is the CompactionCoordinator RPC
service but future services could be converted to use Async processors
as well. This RPC is not using multiplexing as Async multiplexing is an
open issue being worked on in THRIFT-2427. Once that issue is resolve
the plan will be to convert to an async multiplexed processor.

A second thrift service on another port was created in the Manager
for handling async processors because they are handled differently
than sync and you can't combine both sync and async processors
when multiplexing. The coordinator rpc service now advertises the async
port so clients connect to the correct service.

This closes #4664