Add fine-grained options to GrepIterator
Update the previously merged feature to match on column visibilities
using the GrepIterator, which was added in PR #4468

Now, GrepIterator supports a number of fine-grained options to choose
exactly which fields to match on. It preserves the behavior prior to the
changes in PR #4468 by default for the GrepIterator, but retains the
change to the GrepCommand for the shell by enabling the new option to
match on column visibilities.