adds immutable implementation of byte sequence
With this new class there will two implementations of ByteSequence
in the Accumulo API, one being mutable and other immutable. The
ByteSequence abstract class easily supports an immutable implementation
with its current public methods and requires no changes to its public
methods.

There is a single example of where this new class is used in VisibilityFilter
where an iterator caches some data that it does not expect to change for the
lifetime of the class. Using the new immutable byte sequence this expectation
can be enforced.  In a follow on PR could analyze the codebase for other
places where this new class could be used.

New test were written that achieve 100% coverage of both ByteSequence
implementations. Tests were written for some existing code that this
PR did not change.