Modified UserFateStore.create to behave like MetaFateStore (#4787)
MetaFateStore.create will retry forever when a collision happens
when trying to create a Fate transaction. The probabiliy of a
collision is low due to a random UUID being used. Before this change
the UserFateStore.create method would retry 5 times then throw an
exception. This was removed in favor of an unlimited retry so that
the behavior of the two Fate stores is the same.

Closes #4246