Do not fall back to tablet servers when no scan servers by default
By default when using eventual scans and there are no scan servers present
then the scan will fall back to tablet servers.  This could be very suprising
and disruptive when tablet servers do not have much resources allocated for
scanning and scan servers do.  In the case were scan servers are temporarily
down it could cause a thundering herd that overwhelms tablet servers.  This
commit changes the default behavior to not fall back to tablet servers.

This change is targeting 3.1 in order to avoid changing default behavior
in 2.1.X.  The capability to not fall back to tablet servers was introduced
in 2.1.3 via #4671 with a default behavior to fallback.